+++ src/bindings.cpp
@@ -184,7 +184,7 @@ setDartEventCallback(dartVoiceEndedCallback_t voice_ended_callback,
 
 /// Check if the libopus and libogg are available at build time.
 FFI_PLUGIN_EXPORT bool areOpusOggLibsAvailable() {
-#if !defined(NO_OPUS_OGG_LIBS)
+#if !defined(NO_XIPH_LIBS)
   return true;
 #else
   return false;